FrFTL: Added the support of 256MB flash and fixed a bug in trim that will cause unnecessary erases and writes.

+++ radio/src/drivers/frftl.h
@@ -61,7 +61,7 @@ typedef struct {
   uint32_t memoryUsed;
 } FrFTL;
 
-bool ftlInit(FrFTL* ftl, const FrFTLOps* cb, uint8_t flashSizeInMB);
+bool ftlInit(FrFTL* ftl, const FrFTLOps* cb, uint16_t flashSizeInMB);
 void ftlDeInit(FrFTL* ftl);
 
 bool ftlWrite(FrFTL* ftl, uint32_t startSectorNo, uint32_t noOfSectors, const uint8_t* buf);
